requirements.txt
asgiref==3.9.0
beautifulsoup4==4.13.4
Django==5.2.4
et_xmlfile==2.0.0
gunicorn==23.0.0
lxml==6.0.0
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
pytz==2025.2
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
whitenoise==6.9.0
